+++ src/cli/initOpenApiConfig/Types.ts
@@ -1,5 +1,8 @@
+export type RequestScaffoldFormat = 'transport' | 'executor' | 'adapter';
+
 export type CLITemplates = {
     config: Handlebars.TemplateDelegate;
     request: Handlebars.TemplateDelegate;
     requestExecutor: Handlebars.TemplateDelegate;
+    createExecutorAdapter: Handlebars.TemplateDelegate;
 };