+++ src/core/types/base/Templates.model.ts
@@ -28,6 +28,7 @@ export interface Templates {
         httpStatusCode: Handlebars.TemplateDelegate;
         requestExecutor: Handlebars.TemplateDelegate;
         createExecutorAdapter: Handlebars.TemplateDelegate;
+        legacyRequestAdapter: Handlebars.TemplateDelegate;
         interceptors: Handlebars.TemplateDelegate;
         apiErrorInterceptor: Handlebars.TemplateDelegate;
         withInterceptors: Handlebars.TemplateDelegate;