+++ src/templatesCompiled/client/indexCore.ts
@@ -12,7 +12,7 @@ export default {"1":function(container,depth0,helpers,partials,data) {
         return undefined
     };
 
-  return ((stack1 = lookupProperty(helpers,"equals").call(depth0 != null ? depth0 : (container.nullContext || {}),lookupProperty(lookupProperty(data,"root"),"modelsMode"),"classes",{"name":"equals","hash":{},"fn":container.program(2, data, 0),"inverse":container.noop,"data":data,"loc":{"start":{"line":10,"column":0},"end":{"line":13,"column":11}}})) != null ? stack1 : "");
+  return ((stack1 = lookupProperty(helpers,"equals").call(depth0 != null ? depth0 : (container.nullContext || {}),lookupProperty(lookupProperty(data,"root"),"modelsMode"),"classes",{"name":"equals","hash":{},"fn":container.program(2, data, 0),"inverse":container.noop,"data":data,"loc":{"start":{"line":11,"column":0},"end":{"line":14,"column":11}}})) != null ? stack1 : "");
 },"2":function(container,depth0,helpers,partials,data) {
     return "export { BaseDto } from './BaseDto';\nexport { fromArray } from './dtoUtils';\n";
 },"4":function(container,depth0,helpers,partials,data) {
@@ -26,8 +26,8 @@ export default {"1":function(container,depth0,helpers,partials,data) {
     };
 
   return ((stack1 = container.invokePartial(lookupProperty(partials,"header"),depth0,{"name":"header","data":data,"helpers":helpers,"partials":partials,"decorators":container.decorators})) != null ? stack1 : "")
-    + "\nexport { ApiError } from './ApiError';\nexport { RequestConfig, RequestExecutor } from './executor/requestExecutor';\nexport { createExecutorAdapter } from './executor/createExecutorAdapter';\nexport { RequestInterceptor, ResponseInterceptor, ErrorInterceptor } from './interceptors/interceptors';\nexport { apiErrorInterceptor } from './interceptors/apiErrorInterceptor';\nexport { withInterceptors } from './interceptors/withInterceptors';\n"
-    + ((stack1 = lookupProperty(helpers,"if").call(alias1,lookupProperty(lookupProperty(data,"root"),"modelsMode"),{"name":"if","hash":{},"fn":container.program(1, data, 0),"inverse":container.noop,"data":data,"loc":{"start":{"line":9,"column":0},"end":{"line":14,"column":7}}})) != null ? stack1 : "")
-    + ((stack1 = lookupProperty(helpers,"if").call(alias1,lookupProperty(lookupProperty(data,"root"),"useCancelableRequest"),{"name":"if","hash":{},"fn":container.program(4, data, 0),"inverse":container.noop,"data":data,"loc":{"start":{"line":15,"column":0},"end":{"line":17,"column":7}}})) != null ? stack1 : "")
+    + "\nexport { ApiError } from './ApiError';\nexport type { RequestConfig, RequestExecutor } from './executor/requestExecutor';\nexport { createExecutorAdapter } from './executor/createExecutorAdapter';\nexport { createLegacyRequestAdapter } from './executor/legacyRequestAdapter';\nexport { RequestInterceptor, ResponseInterceptor, ErrorInterceptor, RequestRecovery } from './interceptors/interceptors';\nexport { apiErrorInterceptor } from './interceptors/apiErrorInterceptor';\nexport { withInterceptors } from './interceptors/withInterceptors';\n"
+    + ((stack1 = lookupProperty(helpers,"if").call(alias1,lookupProperty(lookupProperty(data,"root"),"modelsMode"),{"name":"if","hash":{},"fn":container.program(1, data, 0),"inverse":container.noop,"data":data,"loc":{"start":{"line":10,"column":0},"end":{"line":15,"column":7}}})) != null ? stack1 : "")
+    + ((stack1 = lookupProperty(helpers,"if").call(alias1,lookupProperty(lookupProperty(data,"root"),"useCancelableRequest"),{"name":"if","hash":{},"fn":container.program(4, data, 0),"inverse":container.noop,"data":data,"loc":{"start":{"line":16,"column":0},"end":{"line":18,"column":7}}})) != null ? stack1 : "")
     + "export { OpenAPI } from './OpenAPI';\nexport type { TOpenAPIConfig } from './OpenAPI';\n";
 },"usePartial":true,"useData":true}